manage context moved to contexts/main
