+++ src/renderer/Providers.tsx
@@ -11,7 +11,7 @@ import { AddressesProvider } from '@/renderer/contexts/main/Addresses';
 import { BootstrappingProvider } from '@app/contexts/main/Bootstrapping';
 import { ChainsProvider } from '@/renderer/contexts/main/Chains';
 import { EventsProvider } from '@/renderer/contexts/main/Events';
-import { ManageProvider } from './screens/Home/Manage/provider';
+import { ManageProvider } from '@/renderer/contexts/main/Manage';
 import { SubscriptionsProvider } from '@app/contexts/main/Subscriptions';
 import { IntervalSubscriptionsProvider } from './contexts/main/IntervalSubscriptions';
